Interpret refs/open-pull/123/merge as pull request
Github is not removing the merge branches of closed prs. Those branches are even updated when master advances. This causes closed prs to be built as if they are open. To avoid this, I created a program that uses the github api to find out which prs are open and create branches only for those. These branches are under refs/open-pull/... since refs/pull/... is read only.

This commit makes the plugin able to find those branches and interpret them as ordinary pull-requests branches so that the status can be reported back to github.